% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnaSeqTutorials-package.R
\docType{package}
\name{RnaSeqTutorials package}
\alias{RnaSeqTutorials package}
\alias{RnaSeqTutorials-package}
\title{RnaSeqTutorials package}
\description{
A simple description of the RnaSeqTutorials package
}
\section{Tutorials}{

There are six tutorials in this package:
\itemize{
\item\code{01_data_science_intro} an introduction to data science using the palmerpenguins package
\item\code{02_exploratory_data_analysis} the exploratory data analysis of a public RNA-Seq dataset
\item\code{03_differential_gene_expression} the differential expression DE analysis of the same dataset
\item\code{04_gene_set_enrichment_analysis} gene set enrichment analysis (GSEA) on the DE gene sets
\item\code{05_ensemble_gene_set_enrichment_analysis} ensemble GSEA of the DE gene sets and comparison to GSEA
\item\code{06_differential_transcript_usage} the differential transcript usage analysis of the same dataset
\item\code{07_non_model_organism_exploratory_data_analysis} the exploratory data analysis of a srpuce RNA-Seq dataset (a non-model organism)
}
}

\examples{
	\dontrun{
	learnr::run_tutorial("01_data_science_intro", package = "RnaSeqTutorials")
	}

}
\seealso{
The vignette
}
\author{
Nicolas Delhomme [aut,cre]
}
\keyword{package}
